+++ admin/intro.php
@@ -27,10 +27,10 @@ use Phyxo\DBLayer\DBLayer;
 // +-----------------------------------------------------------------------+
 
 // Check for upgrade : code inspired from punbb
-if (isset($_GET['action']) and 'check_upgrade' == $_GET['action']) {
+if (isset($_GET['action']) && 'check_upgrade' == $_GET['action']) {
     try {
         $client = new Client();
-        $response = $client->request('GET', PHPWG_URL . '/download/');
+        $response = $client->request('GET', PHYXO_UPDATE_URL);
         if ($response->getStatusCode() == 200 && $response->getBody()->isReadable()) {
             $versions = json_decode($response->getBody(), true);
             $latest_version = $versions[0]['version'];
@@ -41,7 +41,7 @@ if (isset($_GET['action']) and 'check_upgrade' == $_GET['action']) {
         if (preg_match('/.*-dev$/', PHPWG_VERSION, $matches)) {
             $page['infos'][] = \Phyxo\Functions\Language::l10n('You are running on development sources, no check possible.');
         } elseif (version_compare(PHPWG_VERSION, $latest_version) < 0) {
-            $page['infos'][] = \Phyxo\Functions\Language::l10n('A new version of Phyxo is available.');
+            $page['infos'][] = '<a href="./index.php?page=updates">' . \Phyxo\Functions\Language::l10n('A new version of Phyxo is available.') . '</a>';
         } else {
             $page['infos'][] = \Phyxo\Functions\Language::l10n('You are running the latest version of Phyxo.');
         }