+++ admin/theme/js/intro.js
@@ -10,23 +10,25 @@ $(function () {
 				if (data['stat'] != 'ok') {
 					return;
 				}
+				$alertBox = $('section[role="content"] .alert.hide');
 
 				phyxo_update = data.result.phyxo_need_update;
 				ext_update = data.result.ext_need_update;
-				if (!$(".warnings").is('div')) {
-					if (phyxo_update || ext_update) {
-						$("#content").prepend('<div class="warnings"><i class="eiw-icon fa-exclamation"></i><ul></ul></div>');
-						if (phyxo_update) {
-							$(".warnings ul").append('<li>' + phyxo_need_update_msg + '</li>');
-						}
-						if (ext_update) {
-							$(".warnings ul").append('<li>' + ext_need_update_msg + '</li>');
-						}
-					} else {
-						$("#content").prepend('<div class="warnings"><i class="eiw-icon fa-exclamation"></i><ul></ul></div>');
-						$(".warnings ul").append('<li>' + phyxo_is_uptodate_msg + '</li>');
+				if (phyxo_update || ext_update) {
+					$alertBox.addClass('alert-warning');
+					if (ext_update) {
+						$alertBox.prepend('<p>' + ext_need_update_msg + '</p>');
 					}
+					if (phyxo_update) {
+						$alertBox.prepend('<p>' + phyxo_need_update_msg + '</p>');
+					}
+				} else {
+					$alertBox
+						.addClass('alert-success')
+						.prepend('<p>' + phyxo_is_uptodate_msg + '</p>');
 				}
+				$alertBox.removeClass('hide').addClass('show');
+				$('#check-upgrade').parent().remove();
 			}
 		});
 		e.preventDefault();