+++ src/Controller/AdminController.php
@@ -11,6 +11,8 @@
 
 namespace App\Controller;
 
+use App\Repository\UpgradeRepository;
+use Phyxo\DBLayer\iDBLayer;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
@@ -31,6 +33,41 @@ class AdminController extends Controller
 {
     protected $csrfTokenManager, $passwordEncoder, $tagMapper, $commentMapper, $userMapper, $categoryMapper, $rateMapper, $searchMapper, $em, $router, $image_std_params;
     protected $pemURL;
+
+    public function upgrade(iDBLayer $conn)
+    {
+        $tables = $conn->db_get_tables($conn->getPrefix());
+        $columns_of = $conn->db_get_columns_of($tables);
+
+        $result = (new UpgradeRepository($conn))->findAll();
+        $applied_upgrades = $conn->result2array($result, null, 'id');
+
+        if (!in_array(142, $applied_upgrades)) {
+            $current_release = '1.0.0';
+        } elseif (!in_array(144, $applied_upgrades)) {
+            $current_release = '1.1.0';
+        } elseif (!in_array(145, $applied_upgrades)) {
+            $current_release = '1.2.0';
+        } elseif (in_array('validated', $columns_of[$conn->getPrefix() . 'tags'])) {
+            $current_release = '1.3.0';
+        } elseif (!in_array(146, $applied_upgrades)) {
+            $current_release = '1.5.0';
+        } elseif (!in_array(147, $applied_upgrades)) {
+            $current_release = '1.6.0';
+        } elseif (!is_dir(__DIR__ . '/../src/LegacyPages')) {
+            $current_release = '1.8.0';
+        } else {
+            $current_release = '1.9.0';
+        }
+
+        if (\Phyxo\Functions\Upgrade::check_upgrade() && file_exists($upgrade_file = __DIR__ . '/../install/upgrade_' . $current_release . '.php')) {
+            ob_start();
+            include($upgrade_file);
+            ob_end_clean();
+        }
+
+        return $this->redirectToRoute('admin_home');
+    }
 
     public function index(string $legacyBaseDir, Request $request, CsrfTokenManagerInterface $csrfTokenManager, UserPasswordEncoderInterface $passwordEncoder,
                         TagMapper $tagMapper, CommentMapper $commentMapper, UserMapper $userMapper, CategoryMapper $categoryMapper, RateMapper $rateMapper, EntityManager $em,